Refactored files and directories
