+++ android/local.properties
@@ -2,4 +2,4 @@ sdk.dir=/home/ashutosh/Android/Sdk
 flutter.sdk=/home/ashutosh/flutter
 flutter.versionName=1.0.0
 flutter.versionCode=1
-flutter.buildMode=debug
+flutter.buildMode=release
