--- lib/history.dart
@@ -1,61 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:udhari/borrowHistory.dart';
-import 'package:udhari/lendHistory.dart';
-
-class History extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: "Udhari",
-      theme: ThemeData(
-        primarySwatch: Colors.blue,
-      ),
-      debugShowCheckedModeBanner: false,
-      home: HomePage(),
-    );
-  }
-}
-
-class HomePage extends StatefulWidget {
-  @override
-  _HomePageState createState() => _HomePageState();
-}
-
-class _HomePageState extends State<HomePage>
-    with SingleTickerProviderStateMixin {
-  TabController _tabController;
-
-  @override
-  void initState() {
-    super.initState();
-    _tabController = new TabController(vsync: this, initialIndex: 0, length: 4);
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        leading: Icon(Icons.face),
-        title: Text("Udhari History"),
-        bottom: TabBar(
-          controller: _tabController,
-          tabs: <Widget>[
-            Tab(
-              child: Text("Borrow History"),
-            ),
-            Tab(
-              child: Text("Lend History"),
-            ),
-          ],
-        ),
-      ),
-      body: TabBarView(
-        controller: _tabController,
-        children: <Widget>[
-          BorrowHistory(),
-          LendHistory(),
-        ],
-      ),
-    );
-  }
-}
